+++ test/switchboard/native/ArbitrumL2Switchboard.t.sol
@@ -149,7 +149,7 @@ contract ArbitrumL2SwitchboardTest is Setup {
             switchBoardAddress_,
             DEFAULT_BATCH_LENGTH,
             uint32(remoteChainSlug_),
-            uint32(capacitorType_)
+            capacitorType_
         );
 
         scc_.siblingChainSlug = remoteChainSlug_;
